requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.11.17
channels==4.0.0
charset-normalizer==3.3.2
Django==5.0
django-simpleui==2023.12.12
idna==3.6
lxml==4.9.3
numpy==1.26.2
pandas==2.1.4
PyMySQL==1.1.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
simplejson==3.19.2
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tushare==1.2.89
tzdata==2023.3
urllib3==2.1.0
websocket-client==0.57.0
